requirements.txt
typer>=0.9.0,<1
pyyaml>=6.0.1,<7
PyGithub>=2.1.1,<3
loguru>=0.7.2,<1
